+++ jdaviz/core/loaders/importers/spectrum1d/spectrum1d.vue
@@ -1,15 +1,25 @@
 <template>
   <v-container>
     <plugin-select
-      :items="extension_items.map(i => i.label)"
+      :items="extension_items"
       :selected.sync="extension_selected"
       :show_if_single_entry="true"
-      :multiselect="false"
+      :multiselect="multiselect"
+      :exists_in_dc="existing_data_in_dc"
       label="Extension"
-      api_hint="ldr.extension ="
+      api_hint="ldr.importer.extension ="
       :api_hints_enabled="api_hints_enabled"
       hint="Extension to use for the data."
     ></plugin-select>
+
+    <plugin-switch
+      v-if="multiselect && extension_selected.length > 1"
+      :value.sync="concatenate"
+      label="Concatenate"
+      api_hint="ldr.importer.concatenate ="
+      :api_hints_enabled="api_hints_enabled"
+      hint="Concatenate multiple selected spectra into a single spectrum."
+    ></plugin-switch>
 
     <plugin-auto-label
       :value.sync="data_label_value"
@@ -19,9 +29,20 @@
       label="Data Label"
       api_hint="ldr.importer.data_label ="
       :api_hints_enabled="api_hints_enabled"
-      hint="Label to assign to the new data entry."
+      :hint="data_label_is_prefix ? 'Prefix to assign to the new data entry.  Will resolve to the following data labels:' : 'Label to assign to the new data entry.'"
     >
     </plugin-auto-label>
+    <v-row v-if="data_label_is_prefix">
+        <v-chip
+          v-for="suff in data_label_suffices"
+          outlined
+          label
+          :key="suff"
+          style="margin: 4px"
+        >
+          {{data_label_value}}{{suff}}
+        </v-chip>
+    </v-row>
 
     <plugin-viewer-create-new
       :items="viewer_items"
@@ -38,13 +59,13 @@
       api_hint="ldr.importer.viewer ="
       :api_hints_enabled="api_hints_enabled"
       :show_if_single_entry="true"
-      hint="Select the viewer to use for the new data entry."
+      hint="Select the viewer to use for the new data."
     ></plugin-viewer-create-new>
 
     <v-row justify="end">
       <plugin-action-button
         :spinner="import_spinner"
-        :disabled="import_disabled"
+        :disabled="import_disabled || extension_selected.length === 0"
         :results_isolated_to_plugin="false"
         :api_hints_enabled="api_hints_enabled"
         @click="import_clicked">