+++ jdaviz/core/loaders/importers/spectrum_list/__init__.py
@@ -1,1 +1,0 @@
-from .spectrum_list import *  # noqa